+++ tests/testthat/test_datesManipulation.R
@@ -50,51 +50,7 @@ test_that("Private function: identifyDatesFormats ",
 		})
 
 
-## dateDiffs
-#-----------
-dataSet <- data.table(ID = 1:100, 
-                      date1 = seq(from = as.Date("2010-01-01"), 
-                                  to = as.Date("2015-01-01"), 
-                                  length.out = 100), 
-                      date2 = seq(from = as.Date("1910-01-01"), 
-                                  to = as.Date("2000-01-01"), 
-                                  length.out = 100)
-)
 
-# Now let's compute
-dataSet <- diffDates(dataSet, analysisDate = as.Date("2016-11-14"))
-
-test_that("dateDiffs: ",
-          {
-            expect_equal(ncol(dataSet), 6)
-            expect_equal(sum(is.na(dataSet)), 0)
-          })
-		  
-		  
-dataSet <- data.table(ID = 1:100, 
-                      date1 = seq(from = as.Date("2010-01-01"), 
-                                  to = as.Date("2015-01-01"), 
-                                  length.out = 100), 
-                      date2 = seq(from = as.Date("1910-01-01"), 
-                                  to = as.Date("2000-01-01"), 
-                                  length.out = 100)
-)
-test_that("dateDiffs: errors",
-		{
-		expect_error(diffDates(dataSet, analysisDate = "2017-01-01"),  "analysisDate must be a Date")
-		})
-		
-## diffTime
-#----------
-
-test_that("diffTime: ",
-		{
-		expect_equal(diffTime(as.Date("2017-01-02"), as.Date("2017-01-01"), units = "days"), 1)
-		expect_equal(diffTime(as.Date("2017-01-02"), as.Date("2017-01-01"), units = "hours"), 24)
-		expect_equal(diffTime(as.Date("2017-01-02"), as.Date("2017-01-01"), units = "mins"), 1440)
-		expect_equal(diffTime(as.Date("2017-01-02"), as.Date("2017-01-01"), units = "years"), 1 / 365.25)
-		expect_error(diffTime(as.Date("2017-01-02"), as.Date("2017-01-01"), units = "zadtfrey"), "Sorry this unit hasn't been implemented yet")
-		})
 
 ## dateFormatUnifier
 #-------------------